Update breakouttutorialv3.p8
trying to add in different brick functionality. my thinking was to add another table in that keeps different values of powers, that way you can randomly assign a brick with that value and it will be a different brick. Might be easier to create a seperate set of bricks and independently call them each level